update operation implementd on the MyToys page

+++ src/Pages/MyToys/MyToys.jsx
@@ -1,30 +1,45 @@
 /* eslint-disable react/jsx-key */
 import { useContext, useEffect, useState } from "react";
 import { AuthContext } from "../../Provider/AuthProvider";
+import Swal from "sweetalert2";
+import { Link } from "react-router-dom";
 
 const MyToys = () => {
-    const {user} = useContext(AuthContext)
-    const [toys, setToys] = useState([])
-    useEffect(()=>{
-        fetch(`http://localhost:5000/myToys/${user?.email}`)
-        .then(res => res.json())
-        .then(data => setToys(data))
-    },[user?.email])
-    const handleDeleteToy = (id) =>{
-        fetch(`http://localhost:5000/singleToys/${id}`,{
-            method : "DELETE"
+  const { user } = useContext(AuthContext);
+  const [toys, setToys] = useState([]);
+  useEffect(() => {
+    fetch(`http://localhost:5000/myToys/${user?.email}`)
+      .then((res) => res.json())
+      .then((data) => setToys(data));
+  }, [user?.email]);
+  const handleDeleteToy = (id) => {
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`http://localhost:5000/singleToys/${id}`, {
+          method: "DELETE",
         })
-        .then(res => res.json())
-        .then(data =>{
-            console.log(data);
-            const remaining = toys.filter(toy => toy._id !== id)
-            setToys(remaining)
-            alert("successfully deleted")
-        })
-    }
-    return (
-        <div>
-            <div className="overflow-x-auto lg:w-full">
+          .then((res) => res.json())
+          .then((data) => {
+            if (data.deletedCount > 0) {
+              Swal.fire("Deleted!", "Your file has been deleted.", "success");
+            }
+            const remaining = toys.filter((toy) => toy._id !== id);
+            setToys(remaining);
+          });
+      }
+    });
+  };
+  return (
+    <div>
+      <div className="overflow-x-auto lg:w-full">
         <table className="table w-full">
           <thead>
             <tr>
@@ -47,16 +62,23 @@ const MyToys = () => {
                 <td>{toy.price}</td>
                 <td>{toy.quantity}</td>
                 <td className="flex lg:flex-row flex-col gap-4">
-                    <button className="btn d-btn">Update</button>
-                    <button onClick={()=> handleDeleteToy(toy._id)} className="btn d-btn">Delete</button>
+                  <button className="btn d-btn">
+                    <Link to={`/toys/${toy._id}`}>Update</Link>
+                  </button> 
+                  <button
+                    onClick={() => handleDeleteToy(toy._id)}
+                    className="btn d-btn"
+                  >
+                    Delete
+                  </button>
                 </td>
               </tr>
             ))}
           </tbody>
         </table>
       </div>
-        </div>
-    );
+    </div>
+  );
 };
 
 export default MyToys;
